+++ op/console/src/main/java/org/jboss/hal/op/configuration/ConfigurationRoutes.java
@@ -2,6 +2,5 @@ package org.jboss.hal.op.configuration;
 
 public interface ConfigurationRoutes {
 
-    String FINDER_ROUTE = "/configuration/:finderPath?";
     String RESOURCE_ROUTE = "/configuration/resource/:address";
 }