+++ op/console/src/main/java/org/jboss/hal/op/configuration/PathColumn.java
@@ -22,7 +22,6 @@ import org.jboss.hal.op.finder.ColumnProvider;
 import org.patternfly.extension.finder.FinderColumn;
 
 import static java.util.Arrays.asList;
-import static org.jboss.hal.op.configuration.ConfigurationRoutes.RESOURCE_ROUTE;
 import static org.jboss.hal.ui.brick.FinderBricks.crudColumn;
 
 @Dependent
@@ -38,7 +37,7 @@ public class PathColumn implements ColumnProvider {
 
     @Override
     public FinderColumn get() {
-        return crudColumn(ID, "Path", RESOURCE_ROUTE, asList("path", "read-only", "relative-to"),
+        return crudColumn(ID, "Path", asList("path", "read-only", "relative-to"),
                 __ -> TEMPLATE, null);
     }
 }