+++ ui/src/main/java/org/jboss/hal/ui/resource/view/CapabilityReference.java
@@ -219,8 +219,7 @@ class CapabilityReference implements IsElement<HTMLElement>, Attachable {
 
     private void onClick() {
         if (state == State.ONE_RESOURCE && singleTemplate != null) {
-            // TODO [Finding 2] Use ModelBrowserContext.inScope() for link styling
-            //  and context.navigate() instead of SelectInTree.dispatch()
+            // TODO [Finding 2] Use ModelBrowser.inScope() for link styling
             SelectInTree.dispatch(element(), singleTemplate);
         } else if (state == State.MULTIPLE_RESOURCES) {
             popper.show(null);
@@ -252,7 +251,7 @@ class CapabilityReference implements IsElement<HTMLElement>, Attachable {
                         String size = String.valueOf(templates.size());
                         SortedMap<Integer, List<AddressTemplate>> ranked = rank(templates);
                         for (List<AddressTemplate> rank : ranked.values()) {
-                            // TODO [Finding 2] Use ModelBrowserContext.navigate() instead of SelectInTree.dispatch()
+                            // TODO [Finding 2] Use ModelBrowser.inScope() for link styling
                             menuList.addItems(rank, tpl -> listItem()
                                     .add(button(tpl.toString()).link().inline()
                                             .onClick((e, btn) -> SelectInTree.dispatch(element(), tpl))));